third_party/WebKit/LayoutTests/http/tests/push_messaging/unregister.html

<!DOCTYPE html>
<html>
<head>
<title>Test PushRegistration.unregister()</title>
<link rel="manifest" href="resources/push_manifest.json">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../serviceworker/resources/test-helpers.js"></script>
</head>
<body>
<script>
async_test(function(test) {
    var workerUrl = 'resources/empty_worker.js';
    var workerScope = 'resources/scope/' + location.pathname;
    var swRegistration;
    var pushRegistration;
    service_worker_unregister_and_register(test, workerUrl, workerScope)
        .then(function(serviceWorkerRegistration) {
            swRegistration = serviceWorkerRegistration;
            return wait_for_state(test, swRegistration.installing, 'activated');
        })
        .then(function() {
            // If running manually, grant permission when prompted.
            if (self.testRunner)
                testRunner.setPushMessagingPermission(location.origin, true);
            return swRegistration.pushManager.register();
        })
        .then(function(registration) {
            pushRegistration = registration;
            assert_inherits(pushRegistration, 'unregister',
                            'unregister() should be exposed on the PushRegistration object');
            assert_equals(typeof(pushRegistration.unregister), 'function',
                          'PushRegistration.unregister() is a function.');
            return pushRegistration.unregister();
        })
        .then(function(unregistration_result) {
            assert_true(unregistration_result,
                        "unregister() called when correctly registered should be fulfilled with true");
            return pushRegistration.unregister();
        })
        .then(function(unregistration_result) {
            assert_false(unregistration_result,
                         "unregister() called a second time should be fulfilled with false");
            return swRegistration.pushManager.getRegistration();
        })
        .then(function(pushRegistration) {
            assert_equals(pushRegistration, null,
                          "After unregistration, there is no more PushRegistration.");
            return service_worker_unregister_and_done(test, workerScope);
        })
        .catch(unreached_rejection(test));
});
</script>
</body>
</html>
